moved specs into source file
